+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 # dev
+- Fix tile origin detection on tiles that are thinner than the buffer size
 - Make density and class map computation optional by setting the corresponding output sub-dirs to "null"
 
 # v1.0.0